+++ application/controllers/home.php
@@ -9,6 +9,14 @@ class Home extends CI_Controller {
 
 	public function index()
 	{
-		$this->load->view('home'); 
+		$articles = $this->articlemodel->getArticles();
+
+        $data = array('title' => 'All Articles',
+                     'base_url' => base_url(),
+                     'articles' => $articles->result());
+        
+        $this->load->view('head', array('title' => 'All Articles')); 
+		$this->load->view('home', $data);
+        $this->load->view('footer');  
 	}
 }